show all variable
